feat: pectra support in consensus containers (#172)
* feat: pectra support in consensus containers

The number of fields in the `BeaconState` container has increased from
28 to 37 as part of the Pectra (Prague + Electra) upgrade. The tree
height has therefore increased from 5 to 6.

In this hard fork, other containers are affected; we don't use them.

Also included is support for Hoodi in `NetworkConstants.sol`.

* refactor(capsule): move vars around

* fix: build

typo

* fix: use correct chain-id

accidentally used activation epoch for pectra

* fix: update timestamp

Do not accept AI suggestions

* fix: chronological order of timestamps

respond to review comment

* respond to AI comments

* doc: clarify with comment >= timestamp

* fix(pectra): add support for 0x02 prefix

The initialization time of a capsule determines whether it supports
solely the 0x1 prefix or solely the 0x2 prefix